+++ pyogrio/_io.pyx
@@ -402,7 +402,7 @@ cdef get_total_bounds(OGRLayerH ogr_layer, int force):
 
     except CPLE_BaseError:
         bounds = None
-    
+
     return bounds
 
 
@@ -654,13 +654,10 @@ cdef validate_feature_range(OGRLayerH ogr_layer, int skip_features=0, int max_fe
     num_features = max_features
 
     if feature_count == 0:
-        name = OGR_L_GetName(ogr_layer)
-        warnings.warn(f"Layer '{name}' does not have any features to read")
         return 0, 0
 
-    # validate skip_features, max_features
     if skip_features < 0 or skip_features >= feature_count:
-        raise ValueError(f"'skip_features' must be between 0 and {feature_count-1}")
+        skip_features = feature_count
 
     if max_features < 0:
         raise ValueError("'max_features' must be >= 0")
@@ -845,6 +842,10 @@ cdef get_features(
     ]
 
     field_data_view = [field_data[field_index][:] for field_index in range(n_fields)]
+
+    if num_features == 0:
+        return fid_data, geometries, field_data
+
     i = 0
     while True:
         try:
@@ -1242,6 +1243,9 @@ def ogr_open_arrow(
     cdef ArrowArrayStream stream
     cdef ArrowSchema schema
 
+    IF CTE_GDAL_VERSION < (3, 6, 0):
+        raise RuntimeError("Need GDAL>=3.6 for Arrow support")
+
     path_b = path.encode('utf-8')
     path_c = path_b
 
@@ -1251,9 +1255,15 @@ def ogr_open_arrow(
     if fids is not None:
         raise ValueError("reading by FID is not supported for Arrow")
 
-    if skip_features or max_features:
+    IF CTE_GDAL_VERSION < (3, 8, 0):
+        if skip_features:
+            raise ValueError(
+                "specifying 'skip_features' is not supported for Arrow for GDAL<3.8.0"
+            )
+
+    if max_features:
         raise ValueError(
-            "specifying 'skip_features' or 'max_features' is not supported for Arrow"
+            "specifying 'max_features' is not supported for Arrow"
         )
 
     if sql is not None and layer is not None:
@@ -1343,13 +1353,15 @@ def ogr_open_arrow(
         # make sure layer is read from beginning
         OGR_L_ResetReading(ogr_layer)
 
-        IF CTE_GDAL_VERSION < (3, 6, 0):
-            raise RuntimeError("Need GDAL>=3.6 for Arrow support")
-
         if not OGR_L_GetArrowStream(ogr_layer, &stream, options):
             raise RuntimeError("Failed to open ArrowArrayStream from Layer")
 
         stream_ptr = <uintptr_t> &stream
+
+        if skip_features:
+            # only supported for GDAL >= 3.8.0; have to do this after getting
+            # the Arrow stream
+            OGR_L_SetNextByIndex(ogr_layer, skip_features)
 
         # stream has to be consumed before the Dataset is closed
         import pyarrow as pa