+++ CHANGES.md
@@ -20,10 +20,20 @@
     intersect the mask geometry (#285). Note: GDAL < 3.8.0 returns features that
     intersect the bounding box of the mask when using the Arrow interface for
     some drivers; this has been fixed in GDAL 3.8.0.
+-   Removed warning when no features are read from the data source (#299).
 
 ### Other changes
 
 -   test suite requires Shapely >= 2.0
+
+-   using `skip_features` greater than the number of features available in a data
+    layer now returns empty arrays for `read` and an empty DataFrame for
+    `read_dataframe` instead of raising a `ValueError` (#282).
+-   enabled `skip_features` and `max_features` for `read_arrow` and
+    `read_dataframe(path, use_arrow=True)`. Note that this incurs overhead
+    because all features up to the next batch size above `max_features` (or size
+    of data layer) will be read prior to slicing out the requested range of
+    features (#282).
 
 ### Bug fixes